% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animals.R
\docType{data}
\name{animals}
\alias{animals}
\title{List of animals}
\format{A character vector of length 593}
\source{
https://a-z-animals.com/animals/
}
\usage{
animals
}
\description{
List of animals scraped from A-Z Animals.
}
\examples{

sing_alphabet(animals)

}
\keyword{datasets}
