% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{List of countries}
\format{A character vector of length 196}
\source{
https://www.britannica.com/topic/list-of-countries-1993160
}
\usage{
countries
}
\description{
List of countries scraped from Britannica.
}
\examples{

sing_alphabet(countries)

}
\keyword{datasets}
