% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fruits.R
\docType{data}
\name{fruits}
\alias{fruits}
\title{List of fruits}
\format{A character vector of length 114}
\source{
https://simple.wikipedia.org/wiki/List_of_fruits
}
\usage{
fruits
}
\description{
List of fruits scraped from Wikipedia.
}
\examples{

sing_alphabet(fruits)

}
\keyword{datasets}
