% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sing_alphabet.R
\name{sing_alphabet}
\alias{sing_alphabet}
\title{Sing the alphabet song}
\usage{
sing_alphabet(wordbag)
}
\arguments{
\item{wordbag}{A character vector of words to sample from.}
}
\description{
Create the alphabet song by randomly sampling words with the appropriate
initials from the wordbag provided.
}
\examples{

sing_alphabet(fruits)
sing_alphabet(animals)
sing_alphabet(vegetables)
sing_alphabet(countries)

}
