% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vegetables.R
\docType{data}
\name{vegetables}
\alias{vegetables}
\title{List of vegetables}
\format{A character vector of length 126.}
\source{
https://simple.wikipedia.org/wiki/List_of_vegetables
}
\usage{
vegetables
}
\description{
List of vegetables scraped from Wikipedia.
}
\examples{

sing_alphabet(vegetables)

}
\keyword{datasets}
